refactor: move frontend API client to lib/api/client
Consolidate project structure by relocating frontend HTTP service
files from /services/ to /lib/api/client/, clearly separating them
from backend utilities in /lib/api/. This resolves the naming
collision between frontend API clients and backend business logic
services.
