+++ components/matches/NewMatchPage.tsx
@@ -16,9 +16,9 @@ import { useEffect, useState, useCallback, useMemo } from "react";
 import { useRouter } from "next/navigation";
 import type { Player } from "@/types/player.types";
 import type { BackendMatchCreatePayload } from "@/types/match.types";
-import { getPlayers } from "../../services/playerService";
-import { findOrCreateTeam } from "../../services/teamService";
-import { createMatch } from "../../services/matchService";
+import { getPlayers } from "@/lib/api/client/playerService";
+import { findOrCreateTeam } from "@/lib/api/client/teamService";
+import { createMatch } from "@/lib/api/client/matchService";
 import { useForm, Controller } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import * as z from "zod";
