+++ components/teams/TeamDetail.tsx
@@ -17,7 +17,10 @@ import TeamStatsCards from "./TeamStatsCards";
 import TeamMatchesSection from "./TeamMatchesSection";
 import TeamLoadingSkeleton from "./TeamLoadingSkeleton";
 import TeamErrorAlert from "./TeamErrorAlert";
-import { getTeamStatistics, getTeamMatches } from "../../services/teamService";
+import {
+  getTeamStatistics,
+  getTeamMatches,
+} from "@/lib/api/client/teamService";
 
 interface TeamDetailProps {
   teamId: number;
