+++ lib/api/client/index.ts
@@ -1,0 +1,4 @@
+// [>]: Barrel export for frontend API client functions.
+export * from "./matchService";
+export * from "./playerService";
+export * from "./teamService";
